ignore advanced element hiding filters
as described in
https://adblockplus.org/development-builds/new-syntax-for-advanced-eleme
nt-hiding-rules

Previously, these filters would be treated as plain string filters, and
added to the trie. Because they often start with a domain name, they
would add new branches to the trie, and getSubstringsOf would have to
recurse through many levels to determine that these weren’t a match. As
a result, removing these makes filter matching around 10% faster.